+++ packages/saber-plugin-google-analytics/saber-browser.js
@@ -1,9 +1,29 @@
 /* eslint-disable */
 // Google analytics integration for Vue.js renderer
-export default function (ctx) {
-  var router = ctx.router
-  if (process.browser && process.env.NODE_ENV === 'production' && __GA_TRACK_ID__) {
-    (function (i, s, o, g, r, a, m) {
+export default function({ router }) {
+  if (
+    process.browser &&
+    process.env.NODE_ENV === 'production' &&
+    __GA_TRACK_ID__
+  ) {
+    function doNotTrackEnabled() {
+      const dntNumber = parseInt(
+        navigator.msDoNotTrack || // Internet Explorer 9 and 10 vendor prefix
+        window.doNotTrack || // IE 11 uses window.doNotTrack
+          navigator.doNotTrack, // W3C
+        10
+      )
+
+      return dntNumber === 1
+    }
+
+    if (doNotTrackEnabled()) {
+      // Respect doNotTrack setting
+      return
+    }
+
+    // prettier-ignore
+    ;(function(i, s, o, g, r, a, m) {
       i.GoogleAnalyticsObject = r
       i[r] = i[r] || function () {
         (i[r].q = i[r].q || []).push(arguments)
@@ -19,7 +39,7 @@ export default function (ctx) {
     ga('create', __GA_TRACK_ID__, 'auto')
     ga('send', 'pageview')
 
-    router.afterEach(function (to) {
+    router.afterEach(to => {
       ga('set', 'page', to.fullPath)
       ga('send', 'pageview')
     })