+++ packages/saber/lib/plugins/watch-config.js
@@ -53,9 +53,9 @@ exports.apply = api => {
         })
 
         const prevConfig = api.config
-        api.setConfig(data, configPath)
+        await api.setConfig(data, configPath)
         checkIfConfigChanged(api.config, prevConfig)
-
+        await api.hooks.onUpdateConfigFile.promise()
         await emit(api.config)
       })
     }