+++ packages/saber/lib/utils/assetsAttribute.js
@@ -1,5 +1,8 @@
 // @ts-check
 
+const { join } = require('path')
+const { slash } = require('saber-utils')
+
 /**
  * Check if it's external resource
  * @param {string} str
@@ -13,14 +16,20 @@ const MARK_GLOBAL_RE = new RegExp(`"${MARK}([^"]+)"`, 'g')
 /**
  * Prefix MARK to asset path
  * @param {{[key: string]: string}} assets
+ * @param {string} cwd
  */
-const prefixAssets = assets => {
+const prefixAssets = (assets, cwd) => {
   /** @type {{[key: string]: string}} */
   const result = {}
   for (const key of Object.keys(assets)) {
     const value = assets[key]
     if (!isExternal(value) && !value.startsWith(MARK)) {
-      result[key] = `${MARK}${value}`
+      const path = value.startsWith('@')
+        ? value
+        : value.startsWith('module:')
+        ? value.slice(7)
+        : slash(join(cwd, value))
+      result[key] = `${MARK}${path}`
     } else {
       result[key] = value
     }