+++ packages/saber/vue-renderer/app/components/ClientOnly.js
@@ -1,0 +1,13 @@
+export default {
+  name: 'ClientOnly',
+  functional: true,
+  render (h, { parent, children }) {
+    if (parent._isMounted) {
+      return children
+    } else {
+      parent.$once('hook:mounted', () => {
+        parent.$forceUpdate()
+      })
+    }
+  }
+}