+++ packages/saber/vue-renderer/app/create-app.js
@@ -5,13 +5,16 @@ import Vue from 'vue'
 import layouts from '#cache/layouts'
 import createRouter from './router'
 import Meta from './vendor/vue-meta'
-import Layout from './LayoutManager.vue'
+import Layout from './components/LayoutManager.vue'
+import ClientOnly from './components/ClientOnly'
 import extendBrowserApi from '#cache/extend-browser-api'
 import injectConfig from './helpers/inject-config'
 import setTransition from './helpers/set-transition'
+import scrollHandler from './helpers/scroll-handler'
 
 Vue.config.productionTip = false
 
+Vue.component(ClientOnly.name, ClientOnly)
 Vue.component(Layout.name, Layout)
 
 Vue.use(Meta, {
@@ -72,6 +75,13 @@ export default context => {
         transition: null
       }
     },
+    mounted() {
+      scrollHandler(
+        this.$router,
+        this.$router.currentRoute,
+        this.$router.currentRoute
+      )
+    },
     render(h) {
       const transition = Object.assign({}, this.transition)
       const listeners = {}
@@ -87,7 +97,9 @@ export default context => {
       })
       const beforeEnter = listeners['before-enter']
       listeners['before-enter'] = el => {
-        this.$emit('trigger-scroll')
+        this.$nextTick(() => {
+          this.$emit('trigger-scroll')
+        })
         beforeEnter && beforeEnter(el)
       }
       const children = [