+++ website/src/saber-browser.js
@@ -1,8 +1,8 @@
 import 'nprogress/nprogress.css'
-import 'prismjs/themes/prism.css'
-import 'typeface-roboto/index.css'
+import 'typeface-source-sans-pro/index.css'
 import '../../packages/saber-highlight-css/default.css'
 import './css/global.css'
+import './css/prism.css'
 import './css/page.css'
 
 export default ({ router, setHead }) => {