+++ packages/saber-highlight-css/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.5](https://github.com/saberland/saber/compare/saber-highlight-css@0.0.4...saber-highlight-css@0.0.5) (2019-05-29)
+
+### Features
+
+- add support to line numbers in markdown ([#217](https://github.com/saberland/saber/issues/217)) ([1561c61](https://github.com/saberland/saber/commit/1561c61)), closes [#178](https://github.com/saberland/saber/issues/178) [#178](https://github.com/saberland/saber/issues/178)
+
+## [0.0.4](https://github.com/egoist/saber/compare/saber-highlight-css@0.0.3...saber-highlight-css@0.0.4) (2019-05-21)
+
+### Bug Fixes
+
+- **deps:** update peerDependencies in plugins ([e1345c5](https://github.com/egoist/saber/commit/e1345c5))
+
 ## [0.0.3](https://github.com/egoist/saber/compare/saber-highlight-css@0.0.2...saber-highlight-css@0.0.3) (2019-05-08)
 
 ### Bug Fixes