+++ packages/saber-plugin-git-modification-time/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.3](https://github.com/egoist/saber/compare/saber-plugin-git-modification-time@0.0.2...saber-plugin-git-modification-time@0.0.3) (2019-05-21)
+
+### Bug Fixes
+
+- **deps:** update peerDependencies in plugins ([e1345c5](https://github.com/egoist/saber/commit/e1345c5))
+
 ## [0.0.2](https://github.com/egoist/saber/compare/saber-plugin-git-modification-time@0.0.1...saber-plugin-git-modification-time@0.0.2) (2019-05-03)
 
 ### Bug Fixes