+++ packages/saber-plugin-pwa/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/saberland/saber/compare/saber-plugin-pwa@0.3.2...saber-plugin-pwa@0.3.3) (2019-05-28)
+
+### Bug Fixes
+
+- **deps:** update snackbar version ([#216](https://github.com/saberland/saber/issues/216)) ([bc7acc0](https://github.com/saberland/saber/commit/bc7acc0))
+
+## [0.3.2](https://github.com/saberland/saber/compare/saber-plugin-pwa@0.3.1...saber-plugin-pwa@0.3.2) (2019-05-24)
+
+### Bug Fixes
+
+- **saber-plugin-pwa:** fix start_url ([48ec5c7](https://github.com/saberland/saber/commit/48ec5c7))
+
+## [0.3.1](https://github.com/egoist/saber/compare/saber-plugin-pwa@0.3.0...saber-plugin-pwa@0.3.1) (2019-05-21)
+
+### Bug Fixes
+
+- **deps:** update peerDependencies in plugins ([e1345c5](https://github.com/egoist/saber/commit/e1345c5))
+
+# [0.3.0](https://github.com/egoist/saber/compare/saber-plugin-pwa@0.2.1...saber-plugin-pwa@0.3.0) (2019-05-11)
+
+### Bug Fixes
+
+- **saber-plugin-pwa:** read manifest.json from static folder ([52c22f5](https://github.com/egoist/saber/commit/52c22f5))
+
+### Features
+
+- **saber-plugin-pwa:** add appleTouchIcon option ([594c1c3](https://github.com/egoist/saber/commit/594c1c3))
+- **saber-plugin-pwa:** support apple-touch-icon ([b3e0f40](https://github.com/egoist/saber/commit/b3e0f40))
+
 ## [0.2.1](https://github.com/egoist/saber/compare/saber-plugin-pwa@0.2.0...saber-plugin-pwa@0.2.1) (2019-05-08)
 
 **Note:** Version bump only for package saber-plugin-pwa