+++ packages/saber/CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.7](https://github.com/saberland/saber/compare/saber@0.6.6...saber@0.6.7) (2019-05-29)
+
+### Bug Fixes
+
+- decode route hash before scrolling ([b496a00](https://github.com/saberland/saber/commit/b496a00))
+- regression in applyPlugin ([9d01d65](https://github.com/saberland/saber/commit/9d01d65))
+
+## [0.6.6](https://github.com/saberland/saber/compare/saber@0.6.5...saber@0.6.6) (2019-05-29)
+
+### Bug Fixes
+
+- make filterPlugins work in saber-node.js ([#221](https://github.com/saberland/saber/issues/221)) ([f3ec588](https://github.com/saberland/saber/commit/f3ec588))
+- Scroll to anchor not working [#213](https://github.com/saberland/saber/issues/213) ([#214](https://github.com/saberland/saber/issues/214)) ([2286ecd](https://github.com/saberland/saber/commit/2286ecd))
+
+### Features
+
+- add support to line numbers in markdown ([#217](https://github.com/saberland/saber/issues/217)) ([1561c61](https://github.com/saberland/saber/commit/1561c61)), closes [#178](https://github.com/saberland/saber/issues/178) [#178](https://github.com/saberland/saber/issues/178)
+
+## [0.6.5](https://github.com/saberland/saber/compare/saber@0.6.4...saber@0.6.5) (2019-05-28)
+
+### Bug Fixes
+
+- **vue-renderer:** default slot might not exist ([91132a2](https://github.com/saberland/saber/commit/91132a2))
+- check for port only in initial setConfig run ([#215](https://github.com/saberland/saber/issues/215)) ([d778282](https://github.com/saberland/saber/commit/d778282))
+- only check port once ([0b861c4](https://github.com/saberland/saber/commit/0b861c4))
+- support beforePlugins hook in saber-node.js ([9b12cd2](https://github.com/saberland/saber/commit/9b12cd2))
+
+### Features
+
+- add beforePlugins hook ([dd018d5](https://github.com/saberland/saber/commit/dd018d5))
+- add onUpdateConfigFile hook ([de8574f](https://github.com/saberland/saber/commit/de8574f))
+
+## [0.6.4](https://github.com/saberland/saber/compare/saber@0.6.3...saber@0.6.4) (2019-05-25)
+
+### Bug Fixes
+
+- **link-plugin:** remove redundant slash ([#211](https://github.com/saberland/saber/issues/211)) ([3be3592](https://github.com/saberland/saber/commit/3be3592)), closes [#205](https://github.com/saberland/saber/issues/205)
+
+### Features
+
+- run saber server on 3001 port if 3000 is unavailable ([#184](https://github.com/saberland/saber/issues/184)) ([d5d5632](https://github.com/saberland/saber/commit/d5d5632)), closes [#182](https://github.com/saberland/saber/issues/182)
+
+## [0.6.3](https://github.com/egoist/saber/compare/saber@0.6.2...saber@0.6.3) (2019-05-21)
+
+### Bug Fixes
+
+- **markdown:** escape html in heading slug ([777ed4a](https://github.com/egoist/saber/commit/777ed4a))
+
+### Features
+
+- add component ClientOnly ([391183c](https://github.com/egoist/saber/commit/391183c))
+
+## [0.6.2](https://github.com/egoist/saber/compare/saber@0.6.1...saber@0.6.2) (2019-05-11)
+
+### Bug Fixes
+
+- **markdown:** code blocks should alway have a class name ([53f9433](https://github.com/egoist/saber/commit/53f9433))
+- **webpack:** do not include path in output filename in production ([faff48a](https://github.com/egoist/saber/commit/faff48a))
+
+## [0.6.1](https://github.com/egoist/saber/compare/saber@0.6.0...saber@0.6.1) (2019-05-11)
+
+### Features
+
+- plugin for generating responsive images ([#150](https://github.com/egoist/saber/issues/150)) ([d2ed81c](https://github.com/egoist/saber/commit/d2ed81c))
+
+# [0.6.0](https://github.com/egoist/saber/compare/saber@0.5.4...saber@0.6.0) (2019-05-11)
+
+### Bug Fixes
+
+- frontmatter asset loading, fixes [#169](https://github.com/egoist/saber/issues/169) ([#170](https://github.com/egoist/saber/issues/170)) ([7ba1114](https://github.com/egoist/saber/commit/7ba1114))
+- never add new properties to page.attributes ([d0ddafe](https://github.com/egoist/saber/commit/d0ddafe))
+- remove pages.extendPageProp ([5cafde8](https://github.com/egoist/saber/commit/5cafde8))
+- switch back to devalue for cyclical references support ([02e976b](https://github.com/egoist/saber/commit/02e976b))
+
+### Features
+
+- prettify output page component ([992a6a3](https://github.com/egoist/saber/commit/992a6a3))
+- remove trailing slash in permalink ([405157f](https://github.com/egoist/saber/commit/405157f))
+
+### BREAKING CHANGES
+
+- `api.pages.extendPageProp` is removed, you should directly mutate `page` object instead
+
 ## [0.5.4](https://github.com/egoist/saber/compare/saber@0.5.3...saber@0.5.4) (2019-05-10)
 
 ### Bug Fixes